requirements.txt
pandas
openpyxl
pyarrow
flask
flask_sqlalchemy